refactor(master): move layout types to neutral namespace
