+++ vda5050_core/CMakeLists.txt
@@ -171,6 +171,25 @@ target_include_directories(vda5050_client
     $<INSTALL_INTERFACE:${CMAKE_INSTALL_INCLUDEDIR}>
 )
 
+# Layout (LIF) — neutral namespace; master depends on it
+file(GLOB_RECURSE layout_srcs "src/layout/*.cpp")
+add_library(vda5050_layout SHARED ${layout_srcs})
+vda5050_configure_target_alias(
+  TARGET vda5050_layout
+  NAMESPACE ${PROJECT_NAME}::
+  EXPORT_NAME layout
+)
+target_link_libraries(vda5050_layout
+  PUBLIC
+    ${PROJECT_NAME}::types
+    ${PROJECT_NAME}::json_utils
+)
+target_include_directories(vda5050_layout
+  PUBLIC
+    $<BUILD_INTERFACE:${CMAKE_CURRENT_SOURCE_DIR}/include>
+    $<INSTALL_INTERFACE:${CMAKE_INSTALL_INCLUDEDIR}>
+)
+
 # Master
 file(GLOB_RECURSE master_srcs "src/master/*.cpp")
 add_library(vda5050_master SHARED ${master_srcs})
@@ -208,6 +227,7 @@ install(
     vda5050_transport
     vda5050_execution
     vda5050_client
+    vda5050_layout
     vda5050_master
   EXPORT export_${PROJECT_NAME}
   LIBRARY DESTINATION lib
@@ -323,9 +343,9 @@ if(BUILD_TESTING)
   ament_add_gmock(test_master_layout
     test/master/test_unit_layout_loader.cpp
   )
-  target_link_libraries(test_master_layout vda5050_core::master)
+  target_link_libraries(test_master_layout vda5050_core::layout)
   target_compile_definitions(test_master_layout PRIVATE
-    VDA5050_CORE__MASTER__SAMPLE_LAYOUT_PATH="${CMAKE_CURRENT_SOURCE_DIR}/test/master/data/sample_layout.json")
+    VDA5050_CORE__SAMPLE_LAYOUT_PATH="${CMAKE_CURRENT_SOURCE_DIR}/test/master/data/sample_layout.json")
 endif()
 
 # Install linter macro
