+++ vda5050_core/include/vda5050_core/layout/vehicle_type_edge_property.hpp
@@ -16,18 +16,18 @@
  * limitations under the License.
  */
 
-#ifndef VDA5050_CORE__MASTER__LAYOUT__VEHICLE_TYPE_EDGE_PROPERTY_HPP_
-#define VDA5050_CORE__MASTER__LAYOUT__VEHICLE_TYPE_EDGE_PROPERTY_HPP_
+#ifndef VDA5050_CORE__LAYOUT__VEHICLE_TYPE_EDGE_PROPERTY_HPP_
+#define VDA5050_CORE__LAYOUT__VEHICLE_TYPE_EDGE_PROPERTY_HPP_
 
 #include <optional>
 #include <string>
 #include <vector>
 
-#include "vda5050_core/master/layout/layout_action.hpp"
-#include "vda5050_core/master/layout/load_restriction.hpp"
-#include "vda5050_core/master/layout/trajectory.hpp"
+#include "vda5050_core/layout/layout_action.hpp"
+#include "vda5050_core/layout/load_restriction.hpp"
+#include "vda5050_core/layout/trajectory.hpp"
 
-namespace vda5050_core::master::layout {
+namespace vda5050_core::layout {
 
 enum class OrientationType
 {
@@ -62,6 +62,6 @@ struct VehicleTypeEdgeProperty
   std::optional<bool> reentry_allowed;
 };
 
-}  // namespace vda5050_core::master::layout
+}  // namespace vda5050_core::layout
 
-#endif  // VDA5050_CORE__MASTER__LAYOUT__VEHICLE_TYPE_EDGE_PROPERTY_HPP_
+#endif  // VDA5050_CORE__LAYOUT__VEHICLE_TYPE_EDGE_PROPERTY_HPP_
